+++ tests/ItemsStorageTest.php
@@ -5,13 +5,8 @@ declare(strict_types=1);
 namespace Yiisoft\Rbac\Php\Tests;
 
 use PHPUnit\Framework\TestCase;
-use RuntimeException;
-use Yiisoft\Files\FileHelper;
 use Yiisoft\Rbac\ItemsStorageInterface;
-use Yiisoft\Rbac\Permission;
-use Yiisoft\Rbac\Php\ConcurrentItemsStorageDecorator;
 use Yiisoft\Rbac\Php\ItemsStorage;
-use Yiisoft\Rbac\Role;
 use Yiisoft\Rbac\Tests\Common\ItemsStorageTestTrait;
 
 final class ItemsStorageTest extends TestCase
@@ -23,251 +18,122 @@ final class ItemsStorageTest extends TestCase
 
     protected function setUp(): void
     {
-        if ($this->name() === 'testFailCreateDirectory' || $this->name() === 'testCreateNestedDirectory') {
-            FileHelper::ensureDirectory($this->getTempDirectory());
-            FileHelper::clearDirectory($this->getTempDirectory());
-        }
-
         $this->traitSetUp();
     }
 
     protected function tearDown(): void
     {
-        if ($this->name() === 'testFailCreateDirectory' || $this->name() === 'testCreateNestedDirectory') {
-            FileHelper::removeDirectory($this->getTempDirectory());
-        }
-
         $this->clearFixturesFiles();
-    }
-
-    public function testFailCreateDirectory(): void
-    {
-        $directory = $this->getTempDirectory() . '/file.txt';
-        touch($directory);
-
-        $storage = new ItemsStorage($directory);
-        $permission = new Permission('createPost');
-
-        $this->expectException(RuntimeException::class);
-        $this->expectExceptionMessage('Failed to create directory "' . $directory . '". mkdir(): File exists');
-        $storage->add($permission);
-    }
-
-    public function testCreateNestedDirectory(): void
-    {
-        $directory = $this->getTempDirectory() . '/test/create/nested/directory';
-
-        $storage = new ItemsStorage($directory);
-        $storage->add(new Permission('createPost'));
-
-        $this->assertFileExists($directory . '/items.php');
     }
 
     public function testGetAllWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getAll());
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getAll());
     }
 
     public function testGetByNamesWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getByNames(['posts.view']));
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getByNames(['posts.view']));
     }
 
     public function testGetWithConcurrency(): void
     {
-        $this->assertNull($this->getEmptyConcurrentItemsStorage()->get('posts.view'));
+        $this->assertNotNull($this->getEmptyConcurrentItemsStorage()->get('posts.view'));
     }
 
     public function testExistsWithConcurrency(): void
     {
-        $this->assertFalse($this->getEmptyConcurrentItemsStorage()->exists('posts.view'));
+        $this->assertTrue($this->getEmptyConcurrentItemsStorage()->exists('posts.view'));
     }
 
     public function testRoleExistsWithConcurrency(): void
     {
-        $this->assertFalse($this->getEmptyConcurrentItemsStorage()->roleExists('posts.viewer'));
-    }
-
-    public function testAddWithConcurrency(): void
-    {
-        $innerTestStorage = new ItemsStorage($this->getDataPath());
-        $testStorage = new ConcurrentItemsStorageDecorator($innerTestStorage);
-        $actionStorage = $this->getItemsStorage();
-
-        $count = count($actionStorage->getAll());
-        $actionStorage->add(new Permission('test1'));
-        $actionStorage->add(new Permission('test2'));
-
-        $testStorage->add(new Permission('test1'));
-        $this->assertCount($count + 2, $innerTestStorage->getAll());
-    }
-
-    public function testUpdateWithConcurrency(): void
-    {
-        $innerTestStorage = new ItemsStorage($this->getDataPath());
-        $testStorage = new ConcurrentItemsStorageDecorator($innerTestStorage);
-        $actionStorage = $this->getItemsStorage();
-
-        $commonUpdatedItem = $actionStorage->get('posts.view')->withName('posts.view1');
-        $actionStorage->update('posts.view', $commonUpdatedItem);
-        $actionStorage->update('posts.create', $actionStorage->get('posts.create')->withName('posts.create1'));
-
-        $testStorage->update('posts.view', $commonUpdatedItem);
-        $this->assertTrue($innerTestStorage->exists('posts.view1'));
-        $this->assertTrue($innerTestStorage->exists('posts.create1'));
-    }
-
-    public function testRemoveWithConcurrency(): void
-    {
-        $innerTestStorage = new ItemsStorage($this->getDataPath());
-        $testStorage = new ConcurrentItemsStorageDecorator($innerTestStorage);
-        $actionStorage = $this->getItemsStorage();
-
-        $count = count($actionStorage->getAll());
-        $actionStorage->remove('posts.view');
-        $actionStorage->remove('posts.create');
-
-        $testStorage->remove('posts.view');
-        $this->assertCount($count - 2, $innerTestStorage->getAll());
+        $this->assertTrue($this->getEmptyConcurrentItemsStorage()->roleExists('posts.viewer'));
     }
 
     public function testGetRolesWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getRoles());
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getRoles());
     }
 
     public function testGetRolesByNamesWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getRolesByNames(['posts.viewer']));
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getRolesByNames(['posts.viewer']));
     }
 
     public function testGetRoleWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getRole('posts.viewer'));
-    }
-
-    public function testClearPermissionsWithConcurrency(): void
-    {
-        $innerTestStorage = new ItemsStorage($this->getDataPath());
-        $testStorage = new ConcurrentItemsStorageDecorator($innerTestStorage);
-        $actionStorage = $this->getItemsStorage();
-
-        $actionStorage->add(new Role('test1'));
-        $rolesCount = count($actionStorage->getRoles());
-
-        $testStorage->clearPermissions();
-
-        $all = $innerTestStorage->getAll();
-        $this->assertCount($rolesCount, $all);
-        $this->assertContainsOnlyInstancesOf(Role::class, $all);
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getRole('posts.viewer'));
     }
 
     public function testGetPermissionsWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getPermissions());
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getPermissions());
     }
 
     public function testGetPermissionsByNamesWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getPermissionsByNames(['posts.view']));
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getPermissionsByNames(['posts.view']));
     }
 
     public function testGetPermissionWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getPermission('posts.view'));
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getPermission('posts.view'));
     }
 
     public function testGetParentsWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getParents('posts.view'));
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getParents('posts.view'));
     }
 
     public function testGetAccessTreeWithConcurrency(): void
     {
-        $this->markTestSkipped();
-
-        $storage = $this->getEmptyConcurrentItemsStorage();
-
-        $this->expectException(RuntimeException::class);
-        $this->expectExceptionMessage('Base item not found.');
-        $storage->getAccessTree('posts.view');
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getAccessTree('posts.view'));
     }
 
     public function testGetDirectChildrenWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getDirectChildren('posts.viewer'));
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getDirectChildren('posts.viewer'));
     }
 
     public function testGetAllChildrenWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getAllChildren('posts.viewer'));
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getAllChildren('posts.viewer'));
     }
 
     public function testGetAllChildRolesWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getAllChildRoles('posts.redactor'));
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getAllChildRoles('posts.redactor'));
     }
 
     public function testGetAllChildPermissionsWithConcurrency(): void
     {
-        $this->assertEmpty($this->getEmptyConcurrentItemsStorage()->getAllChildPermissions('posts.viewer'));
+        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getAllChildPermissions('posts.viewer'));
     }
 
     public function testHasChildrenWithConcurrency(): void
     {
-        $this->assertFalse($this->getEmptyConcurrentItemsStorage()->hasChildren('posts.viewer'));
+        $this->assertTrue($this->getEmptyConcurrentItemsStorage()->hasChildren('posts.viewer'));
     }
 
     public function testHasChildWithConcurrency(): void
     {
-        $this->assertFalse($this->getEmptyConcurrentItemsStorage()->hasChild('posts.viewer', 'posts.view'));
+        $this->assertTrue($this->getEmptyConcurrentItemsStorage()->hasChild('posts.viewer', 'posts.view'));
     }
 
     public function testHasDirectChildWithConcurrency(): void
     {
-        $this->assertFalse($this->getEmptyConcurrentItemsStorage()->hasDirectChild('posts.viewer', 'posts.view'));
-    }
-
-    public function testAddChildWithConcurrency(): void
-    {
-        $innerTestStorage = new ItemsStorage($this->getDataPath());
-        $testStorage = new ConcurrentItemsStorageDecorator($innerTestStorage);
-        $actionStorage = $this->getItemsStorage();
-
-        $actionStorage->addChild('posts.viewer', 'posts.create');
-        $actionStorage->addChild('posts.viewer', 'posts.update');
-
-        $testStorage->addChild('posts.viewer', 'posts.create');
-        $this->assertTrue($innerTestStorage->hasChild('posts.viewer', 'posts.create'));
-        $this->assertTrue($innerTestStorage->hasChild('posts.viewer', 'posts.update'));
-    }
-
-    public function testMultipleCallsWithConcurrency(): void
-    {
-        $testStorage = $this->getItemsStorageForModificationAssertions();
-        $actionStorage = $this->getItemsStorage();
-        $count = count($actionStorage->getAll());
-
-        $actionStorage->add(new Permission('test1'));
-        $this->assertCount($count + 1, $testStorage->getAll());
-
-        $actionStorage->add(new Permission('test2'));
-        $this->assertCount($count + 2, $testStorage->getAll());
+        $this->assertTrue($this->getEmptyConcurrentItemsStorage()->hasDirectChild('posts.viewer', 'posts.view'));
     }
 
     protected function createItemsStorage(): ItemsStorageInterface
     {
-        return new ConcurrentItemsStorageDecorator(new ItemsStorage($this->getDataPath()));
+        return new ItemsStorage($this->getDataPath());
     }
 
     protected function getItemsStorageForModificationAssertions(): ItemsStorageInterface
     {
-        return $this->createItemsStorage();
-    }
-
-    private function getTempDirectory(): string
-    {
-        return __DIR__ . '/temp';
+        return str_ends_with($this->name(), 'WithConcurrency') ? $this->createItemsStorage() : $this->getItemsStorage();
     }
 
     private function getEmptyConcurrentItemsStorage(): ItemsStorageInterface
